requirements.txt
flask
flask-sqlalchemy
werkzeug
requests
python-dotenv 